[REFACTOR]: client-apple token 관련 Refactoring
