+++ clients/client-apple/build.gradle.kts
@@ -1,4 +1,5 @@
 dependencies {
+    implementation(project(":core:core-domain"))
     implementation("org.springframework.boot:spring-boot-starter-web")
     implementation("org.bouncycastle:bcpkix-jdk18on:1.78.1")
 }
