+++ clients/client-apple/src/main/kotlin/com/myongjiway/client/apple/model/ApplePublicKey.kt
@@ -1,0 +1,10 @@
+package com.myongjiway.client.apple.model
+
+data class ApplePublicKey(
+    val kty: String,
+    val kid: String,
+    val use: String,
+    val alg: String,
+    val n: String,
+    val e: String,
+)
