+++ clients/client-apple/src/main/kotlin/com/myongjiway/client/apple/util/AppleJwtUtil.kt
@@ -1,0 +1,9 @@
+package com.myongjiway.client.apple.util
+
+import com.myongjiway.client.apple.AppleClient
+import com.myongjiway.core.domain.token.TokenValidator
+
+class AppleJwtUtil(
+    private val tokenValidator: TokenValidator,
+    private val appleClient: AppleClient,
+)
