+++ clients/client-mqtt/src/main/kotlin/com/myongjiway/client/mqtt/config/MqttClientConfig.kt
@@ -6,7 +6,7 @@ import com.amazonaws.services.iot.client.AWSIotQos
 import com.amazonaws.services.iot.client.AWSIotTopic
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import com.myongjiway.clientmqtt.util.AwsIotUtil
+import com.myongjiway.client.mqtt.util.AwsIotUtil
 import com.myongjiway.core.domain.buslocation.BusLocation
 import com.myongjiway.core.domain.buslocation.BusLocationService
 import org.springframework.beans.factory.annotation.Value
